chore: move demo under packages
use create-react-app for demo
