--- demo/package.json
@@ -1,19 +1,0 @@
-{
-  "name": "cerebral-demo",
-  "version": "1.0.0",
-  "description": "",
-  "main": "main.js",
-  "scripts": {
-    "start": "webpack-dev-server"
-  },
-  "author": "Christian Alfoni <christianalfoni@gmail.com>",
-  "license": "MIT",
-  "private": true,
-  "devDependencies": {
-    "babel-core": "^6.17.0",
-    "babel-loader": "^6.2.5",
-    "html-webpack-plugin": "^2.22.0",
-    "webpack": "^1.13.2",
-    "webpack-dev-server": "^1.16.1"
-  }
-}
