+++ packages/cerebral-demo/package.json
@@ -1,0 +1,24 @@
+{
+  "name": "cerebral-demo",
+  "version": "0.1.0",
+  "private": true,
+  "author": "Christian Alfoni <christianalfoni@gmail.com>",
+  "contributors": [
+    "Aleksey Guryanov <gurianov@gmail.com>"
+  ],
+  "devDependencies": {
+    "react-scripts": "0.6.1"
+  },
+  "dependencies": {
+    "react": "^15.3.2",
+    "react-dom": "^15.3.2",
+    "cerebral": "^2.0.0-0",
+    "cerebral-router": "^1.0.0-0"
+  },
+  "scripts": {
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test --env=jsdom",
+    "eject": "react-scripts eject"
+  }
+}
