+++ packages/cerebral-demo/src/computeds/upperCaseTitle.js
